+++ ContentPatcher/release-notes.md
@@ -1,6 +1,11 @@
 [← back to readme](README.md)
 
 # Release notes
+## 1.5.1
+* Added token support in `EditData` keys.
+* Fixed error setting `EditData` entries to `null` since 1.5.
+* Fixed error using tokens in `Enabled` field since 1.5.
+
 ## 1.5
 * Added support for dynamic tokens defined by the modder.
 * Added new tokens: